Moved Base16Theme type into types.js file
