+++ frontend/Breadcrumb.js
@@ -12,7 +12,7 @@
 
 import type Store from './Store';
 import type {ElementID} from './types';
-import type {Base16Theme} from './Themes/Themes';
+import type {Base16Theme} from './types';
 
 var React = require('react');
 var decorate = require('./decorate');
