+++ frontend/Props.js
@@ -13,7 +13,7 @@
 var React = require('react');
 var PropVal = require('./PropVal');
 
-import type {Base16Theme} from './Themes/Themes';
+import type {Base16Theme} from './types';
 
 class Props extends React.Component {
   context: {
