+++ frontend/SettingsPane.js
@@ -19,7 +19,7 @@ var {PropTypes} = React;
 
 var decorate = require('./decorate');
 
-import type {Base16Theme} from './Themes/Themes';
+import type {Base16Theme} from './types';
 
 type EventLike = {
   keyCode: number,
