+++ frontend/TabbedPane.js
@@ -13,7 +13,7 @@
 var React = require('react');
 var decorate = require('./decorate');
 
-import type {Base16Theme} from './Themes/Themes';
+import type {Base16Theme} from './types';
 
 class TabbedPane extends React.Component {
   context: {
