+++ frontend/detail_pane/DetailPane.js
@@ -13,7 +13,7 @@
 var assign = require('object-assign');
 var React = require('react');
 
-import type {Base16Theme} from '../Themes/Themes';
+import type {Base16Theme} from '../types';
 
 class DetailPane extends React.Component {
   context: {
